V3- 7/11/24
Added more HTML and CSS to the watchlist page to make sure that the form to add new movies to the watchlist looked more appealing to the user and had all the necessary fields and buttons.  I also changed some CSS  in the Home page as i noticed that in the some of the responsive features weren't working properly so I changed the margin so that is would adjust properly.