V6-17/11/24
Added more JavaScript elements to the watchlist page and also worked on a bit of the backend for the watchlist form. I also installed the node modules into the file and worked on the movies page and added the movie cards and added the CSS to make it responsive, the images and titles still need to be changed. I also decided to get rid of the Favourites page as it wasn't really necessary for the PWA.