V7-20/11/24
Some stylistic changes were made in the CSS of the Movies page to make the page fit the aesthetic.  Some changes were also made to the Watchlist HTML and CSS pages with another section called priority being added to the form which allows for the movies to be selected as low, medium, and high priorities, determining when the movie should be watched according to the user.