V8-22/11/24
Worked on the Movies page and made changes to the CSS and HTML pages. The previous card created for the movies were replaced with flip cards that were made responsive. Having the cards as flip cards allowed for there to be space to place the synopsis and by making sure that the cards were responsive it made sure that the cards fit the criteria for the project. The cards were all created however, the content for some of the cards need to added and fixed.