V9-23/11/24
Completed all the card designs for the Movies page using HTML and CSS elements. The Watchlist form on the Watchlist page was also worked on and the server.js and Home.js  were worked on to make the backend work. I also reinstalled the node.js to solve issues with the watchlist items not creating and made sure to check the Watchlist_log.db to make sure that the items added also were shown in the table. Although the form creates the Watchlist item it still needs to allow users to edit and delete the items in the list.