V11-26/11/24
Created the delete button for the Watchlist form through the HTML, CSS and JS files with changes being made to both the Watchlist.js (used to be Home.js but was renamed to be relevant to the connected file). I also added the edit button to the interface however, I still need to add the relevant code so that it works.