V17-18/12/24
Continued to work on finishing the CRUD functions and ended up adding a sort function and finished the edit function. I also worked on making it an actual PWA with the service worker, manifest and icon files being created. I still need to work on making the service worker fully functional and making it work offline.