V8- V9 : 24/6/25 - 27/6/25
V8: Continued to work on making the functionalities for the registration and login page work, however the when the program was run the database for the users would not create so the functionalities wouldn't actually work.

V9: Continued to work on the login and registration functionalities and changed the routing and also adding more code in the LoginandSignup.js page. I was able to get the registration function working properly with it saving to the Users table within Account.db and also made changes within the server,js to make the registration function work. However, I still need to apply bcrypt and saltrounds to the forms and also make the login function redirect properly.