V10-28/6/25
Continued to work on the loginandSignup page. The redirection to the homepage when successfully logged in was completed. Bcrypt and Saltround functionalities were also added to the system and were functional. However, I still need to make sure when the user registers there cannot use the same email or username.