V11-29/6/25
Added the functionality to make sure that the username or email aren't repeated when registering. I added an error message which displays to the user when the register to inform them that the username has been taken. I also ran various checks to ensure that the system was working properly and was not susceptible to the threats such as SQL injection.