V12: 30/6/25 -3/7/25
Completed the limit of attempts functionality for the login form and set a timer for 4 minutes after the attempts have been exhausted, which makes the user unable to log in until the timer runs out. I also added a logout button to all the other pages of the MovieForMe system. I also added more comments throughout the system and also added another condition for the password during registration which is that they also need at least one special character.